split integreation tests
